Add basic layout with dummy data
Added a group of selectors that populate the rest of the page. Ideally
the current hero state would be persisted in the url for easier
copy+pasting but just in memory for now.

Using the first of each style/form/archetype from the book as sample
data. Copying the rest out like that could be tedious - can likely
script something from the sheets version of this ~> csv ~> strings.ts or
similar

there's a few TODO comments, most are a feature wishlist