Move v1 to /v1/ route, add redirect
The app should continue working as before, and the redirect can catch
any old stateful links that might be clicked. v2 is just a simple "WIP"
for now, will add to that shortly.

There's a feature wishlist in the readme as something of a roadmap
as/when I find time