Add hello-world graphql func with typescript
it's not the prettiest implementation - apollo once had a gcp/functions
server but it's in a deprecated version of the library.
It also only serves HTTP. this is enough for actual use but tools like
postman appear to try using the websocket side of things. also the lack
of graphiql and schema fetching will make clientside dev a bit awkward